backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
supabase
python-jose[cryptography]
